replaced runtestredirect with run_shell_cmd
