+++ ChangeLog
@@ -1,9 +1,12 @@
 2015-07-22  Titus Brown  <titus@idyll.org>
 
    * tests/test_streaming_io.py: new shell cmd tests for streaming/piping.
-   * scripts/interleave-reads.py: force exactly two input files.
+   * tests/khmer_tst_utils.py: refactor/replace runtestredirect(...) with
+   scriptpath(...) and run_shell_cmd(...).
    * scripts/test_scripts.py: remove test_interleave_reads_broken_fq_4 for
-   only one input file for interleave-reads.py.
+   only one input file for interleave-reads.py; replace runscriptredirect call
+   with run_shell_cmd.
+   * scripts/interleave-reads.py: force exactly two input files.
    * scripts/split-paired-reads.py: fix print statement; clarify output.
    * scripts/{normalize-by-median.py,sample-reads-randomly.py,
    trim-low-abund.py}: if stdin is supplied for input, check that -o
@@ -17,6 +20,7 @@
    * tests/test-data/paired.fq.2: removed extraneous newline from end.
    * tests/{test_normalize_by_median,test_script_arguments,test_scripts}.py:
    added tests for new code.
+   * scripts/oxli: added script for running tests in development directory.
 
 2015-07-22  Jacob Fenton  <bocajnotnef@gmail.com>
